cam_server/requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
numpy<2.0.0
requests
